Changed transformer script, added 3 new scripts
Changed the transformer to move diagonally and the speed of when it moves along the y axis. The rotation is now faster and on multiple axises. Made the ObjectMover to move object on the x axis. SimpleRotator to rotate an object on the x axis. ScalePulser makes the object grow and shrink smoothly.