Added new scene for Rigidbody and Colliders lab
Created a small scene with a floor and wall. I made a cube and sphere that can fall to the ground and collide with the floor. I then made another cube which can move with the ApplyForce script. I also made a tower of cubes that the moving cube could knock over for fun