Health Bar UI
I added in the script for the HealthContainer (and an empty game object to contain that script), made a HealthData variable and SimpleFloatBehavior which was used on the Spike Balls and an enemy to decrease a health bar when touched.  Changed the code a bit for the FlipTransform script as well because I thought it wasn't working, but in reality I am a dingus and was pressing the wrong keys in play mode. Works fine though, so we ignore that.