Mana Bar UI
Added in a script to add or reduce mana depending on the situation. The strawberry added to the scene adds mana to the bar when touched, and when I hit the V button, the mana decreases. I added a function to the SimpleCharacterController script that called the manacontainter script and image script so that it would actually work.