Attachable Kiwi
Made four scripts. One to make a scriptable object(Id), a simple ID behavior, an ID match behavior, and a script to attach an object to the character. Kiwi has an ID and the fire box has an animation, but the fire still needs to be set up to have the kiwi turn it off.